removing NormalizedEthereumTx in favour of a more idomatic codebase approach
