--- tx/eth_tx_data_test.go
@@ -1,213 +1,0 @@
-// Copyright (c) 2026 The VeChainThor developers
-//
-// Distributed under the GNU Lesser General Public License v3.0 software license, see the accompanying
-// file LICENSE or <https://www.gnu.org/licenses/lgpl-3.0.html>
-
-package tx
-
-import (
-	"math/big"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// TestNewEthereumTransaction_1559 covers the EIP-1559 path.
-func TestNewEthereumTransaction_1559(t *testing.T) {
-	rawBytes, err := defaultEth1559Builder().BuildRaw(ethTestKey)
-	require.NoError(t, err)
-	norm, err := NormalizeEthereumTx(rawBytes, testChainID)
-	require.NoError(t, err)
-
-	trx := NewEthereumTransaction(norm)
-
-	assert.Equal(t, TypeEthTyped1559, trx.Type())
-	assert.Equal(t, norm.Hash, trx.ID())
-
-	origin, err := trx.Origin()
-	require.NoError(t, err)
-	assert.Equal(t, norm.Sender, origin)
-
-	assert.Equal(t, norm.GasLimit, trx.Gas())
-	assert.Equal(t, norm.MaxFeePerGas, trx.MaxFeePerGas())
-	assert.Equal(t, norm.MaxPriorityFeePerGas, trx.MaxPriorityFeePerGas())
-
-	clauses := trx.Clauses()
-	require.Len(t, clauses, 1)
-	assert.Equal(t, norm.To, clauses[0].To())
-	assert.Equal(t, norm.Value, clauses[0].Value())
-	assert.Empty(t, clauses[0].Data())
-
-	// VeChain stubs: chain tag returns 0 (Ethereum txs use chainID for replay protection).
-	assert.Equal(t, byte(0), trx.ChainTag())
-	assert.Equal(t, uint32(0), trx.BlockRef().Number())
-	assert.Equal(t, uint32(maxUint32), trx.Expiration())
-	assert.False(t, trx.IsExpired(uint32(maxUint32)), "Ethereum txs must not expire at any uint32 block number")
-	assert.Nil(t, trx.DependsOn())
-	assert.Equal(t, Features(0), trx.Features(), "no VeChain feature flags on Ethereum txs")
-}
-
-// TestNewEthereumTransaction_1559ContractCreation verifies nil To (contract creation) for EIP-1559.
-func TestNewEthereumTransaction_1559ContractCreation(t *testing.T) {
-	trx, err := defaultEth1559Builder().To(nil).Build(ethTestKey)
-	require.NoError(t, err)
-
-	clauses := trx.Clauses()
-	require.Len(t, clauses, 1)
-	assert.True(t, clauses[0].IsCreatingContract())
-}
-
-// TestNewEthereumTransaction_1559MarshalRoundtrip verifies that MarshalBinary followed by
-// UnmarshalBinary produces a Transaction with identical observable properties.
-func TestNewEthereumTransaction_1559MarshalRoundtrip(t *testing.T) {
-	testData := []byte{0xca, 0xfe, 0xba, 0xbe}
-	original, err := defaultEth1559Builder().Data(testData).Build(ethTestKey)
-	require.NoError(t, err)
-
-	encoded, err := original.MarshalBinary()
-	require.NoError(t, err)
-
-	// Block-body encoding: first byte must be the 0x02 EIP-1559 type byte.
-	require.NotEmpty(t, encoded)
-	assert.Equal(t, TypeEthTyped1559, encoded[0])
-
-	var decoded Transaction
-	require.NoError(t, decoded.UnmarshalBinary(encoded))
-
-	assert.Equal(t, TypeEthTyped1559, decoded.Type())
-	assert.Equal(t, original.ID(), decoded.ID())
-
-	originOrig, err := original.Origin()
-	require.NoError(t, err)
-	originDecoded, err := decoded.Origin()
-	require.NoError(t, err)
-	assert.Equal(t, originOrig, originDecoded)
-
-	assert.Equal(t, original.MaxFeePerGas(), decoded.MaxFeePerGas())
-	assert.Equal(t, original.MaxPriorityFeePerGas(), decoded.MaxPriorityFeePerGas())
-	assert.Equal(t, original.Gas(), decoded.Gas())
-}
-
-// TestNewEthereumTransaction_UnsupportedTypePanics verifies the panic guard for unknown types.
-func TestNewEthereumTransaction_UnsupportedTypePanics(t *testing.T) {
-	norm := &NormalizedEthereumTx{TxType: 0xFF}
-	assert.Panics(t, func() { NewEthereumTransaction(norm) })
-}
-
-// TestNewEthereumTransaction_Hash verifies that Hash() == ID() == ethTxHash for Ethereum txs,
-// preventing the latent bug where unexported fields cause rlp.Encode to produce an empty
-// encoding and every Ethereum tx would share the same Hash().
-func TestNewEthereumTransaction_Hash(t *testing.T) {
-	tx1559, err := defaultEth1559Builder().Build(ethTestKey)
-	require.NoError(t, err)
-	assert.Equal(t, tx1559.ID(), tx1559.Hash(), "Eth1559: Hash must equal ID (ethTxHash)")
-}
-
-// TestNewEthereumTransaction_1559Copy verifies that copy() deep-copies an EIP-1559 tx body.
-func TestNewEthereumTransaction_1559Copy(t *testing.T) {
-	original, err := defaultEth1559Builder().Build(ethTestKey)
-	require.NoError(t, err)
-
-	copied := &Transaction{body: original.body.copy()}
-
-	assert.Equal(t, original.ID(), copied.ID())
-	assert.Equal(t, original.Gas(), copied.Gas())
-	assert.Equal(t, original.MaxFeePerGas(), copied.MaxFeePerGas())
-	assert.Equal(t, original.MaxPriorityFeePerGas(), copied.MaxPriorityFeePerGas())
-
-	// Mutation isolation: mutating the original body must not affect the copy.
-	original.body.(*eth1559TxData).maxFee.SetInt64(0)
-	assert.Equal(t, big.NewInt(10e9), copied.MaxFeePerGas(),
-		"copied body must not share maxFee pointer with original")
-}
-
-// TestNewEthereumTransaction_SetSignaturePanics verifies that calling WithSignature on an
-// Ethereum tx panics with a clear message.
-func TestNewEthereumTransaction_SetSignaturePanics(t *testing.T) {
-	tx1559, err := defaultEth1559Builder().Build(ethTestKey)
-	require.NoError(t, err)
-	assert.Panics(t, func() { tx1559.WithSignature(make([]byte, 65)) })
-}
-
-// TestNewEthereumTransaction_1559EffectiveGasPrice verifies the EIP-1559 formula
-// min(maxFeePerGas, maxPriorityFeePerGas + baseFee) for EthTyped1559.
-// Default builder: maxFeePerGas=10 Gwei, maxPriorityFeePerGas=1 Gwei.
-func TestNewEthereumTransaction_1559EffectiveGasPrice(t *testing.T) {
-	rawBytes, err := defaultEth1559Builder().BuildRaw(ethTestKey)
-	require.NoError(t, err)
-	norm, err := NormalizeEthereumTx(rawBytes, testChainID)
-	require.NoError(t, err)
-	trx := NewEthereumTransaction(norm)
-
-	tests := []struct {
-		name     string
-		baseFee  *big.Int
-		expected *big.Int
-	}{
-		// priority(1e9) + baseFee(2e9) = 3e9 < maxFee(10e9) → uncapped
-		{"uncapped", big.NewInt(2e9), big.NewInt(3e9)},
-		// priority(1e9) + baseFee(12e9) = 13e9 > maxFee(10e9) → fee-capped
-		{"fee_capped", big.NewInt(12e9), big.NewInt(10e9)},
-		// priority(1e9) + baseFee(9e9) = 10e9 == maxFee(10e9) → boundary (capped)
-		{"boundary", big.NewInt(9e9), big.NewInt(10e9)},
-		// priority(1e9) + baseFee(0) = 1e9 → zero base fee
-		{"zero_basefee", big.NewInt(0), big.NewInt(1e9)},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expected, trx.EffectiveGasPrice(tc.baseFee, nil))
-		})
-	}
-}
-
-// TestNewEthereumTransaction_EffectivePriorityFeePerGas verifies
-// min(maxFeePerGas − baseFee, maxPriorityFeePerGas) for EthTyped1559.
-func TestNewEthereumTransaction_EffectivePriorityFeePerGas(t *testing.T) {
-	raw1559, err := defaultEth1559Builder().BuildRaw(ethTestKey)
-	require.NoError(t, err)
-	norm1559, err := NormalizeEthereumTx(raw1559, testChainID)
-	require.NoError(t, err)
-	trx1559 := NewEthereumTransaction(norm1559)
-
-	tests := []struct {
-		name     string
-		baseFee  *big.Int
-		expected *big.Int
-	}{
-		// EthTyped1559: maxFee=10 Gwei, maxPriority=1 Gwei
-		// min(10e9 − 2e9, 1e9) = 1e9 (priority-capped)
-		{"1559_priority_capped", big.NewInt(2e9), big.NewInt(1e9)},
-		// min(10e9 − 9e9, 1e9) = 1e9 (boundary, priority-capped)
-		{"1559_boundary", big.NewInt(9e9), big.NewInt(1e9)},
-		// min(10e9 − 10e9, 1e9) = 0 (maxFee consumed by baseFee)
-		{"1559_zero_priority", big.NewInt(10e9), big.NewInt(0)},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			// legacyTxBaseGasPrice and provedWork are only used for TypeLegacy (VeChain native);
-			// pass nil for both.
-			// Use Cmp instead of Equal: big.Int subtraction producing zero yields
-			// abs:{} (empty slice) rather than abs:nil, which trips reflect.DeepEqual.
-			got := trx1559.EffectivePriorityFeePerGas(tc.baseFee, nil, nil)
-			assert.Equal(t, 0, tc.expected.Cmp(got), "expected %s, got %s", tc.expected, got)
-		})
-	}
-}
-
-// TestTransaction_EthChainID verifies that EthChainID() returns the embedded chain ID
-// for TypeEthTyped1559 and zero for all VeChain-native types.
-func TestTransaction_EthChainID(t *testing.T) {
-	eth1559, err := defaultEth1559Builder().ChainID(99999).Build(ethTestKey)
-	require.NoError(t, err)
-	assert.Equal(t, uint64(99999), eth1559.EthChainID(), "EthTyped1559 must return its embedded chain ID")
-
-	legacy := NewBuilder(TypeLegacy).Build()
-	assert.Equal(t, uint64(0), legacy.EthChainID(), "TypeLegacy must return 0")
-
-	dynFee := NewBuilder(TypeDynamicFee).Build()
-	assert.Equal(t, uint64(0), dynFee.EthChainID(), "TypeDynamicFee must return 0")
-}
-
-// maxUint32 mirrors math.MaxUint32 for use in assertions without an import.
-const maxUint32 = 1<<32 - 1
